Renamed files: `observer` to `neuron`
