+++ src/neurons/system:column/delete-columns.spec.ts
