+++ src/neurons/system:schema/delete-tables.spec.ts
