+++ src/neurons/system:schema/delete-tables.ts
@@ -1,18 +1,14 @@
 import _ from 'lodash';
-import chai from 'chai';
-import { Knex } from 'knex';
 
 // Classes
 import { AutoInstall } from '@classes/autoinstall';
-import { Observer } from '@classes/observer';
+import { Observer } from '@classes/neuron';
 import { Signal } from '@classes/signal';
 import { Record } from '@classes/record';
-import { Schema } from '@classes/schema';
 
 // Typedefs
-import { ObserverRing } from '@typedefs/observer';
+import { ObserverRing } from '@typedefs/neuron';
 import { SchemaType } from '@typedefs/schema';
-
 
 export default class extends Observer {
     toName(): string {
@@ -27,7 +23,7 @@ export default class extends Observer {
         return ObserverRing.Post;
     }
 
-    onCreate(): boolean {
+    onDelete(): boolean {
         return true;
     }
 
@@ -36,20 +32,19 @@ export default class extends Observer {
     }
 
     async one(signal: Signal, record: Record) {
-        // Setup
         let schema_name = record.data.name;
         let schema_type = record.data.type;
         let auto = new AutoInstall(signal.kernel);
 
-        // Only create schemas that are marked as `database` types
+        // Only delete schemas that are marked as `database` types
         if (schema_type !== 'database') {
             return;
         }
 
-        // Create the empty table with no default columns
-        await auto.createTable(schema_name, table => {});
+        // Drop the table
+        await auto.deleteTable(schema_name);
 
-        // Add to kernel metadata
-        signal.kernel.meta.schemas.set(schema_name, Schema.from(record.data));
+        // Remove from kernel metadata
+        signal.kernel.meta.schemas.delete(schema_name);
     }
 }
