+++ src/typedefs/neuron.ts
